pyproject.toml
[project]
name = "personal-ai-employee"
version = "0.1.0"
description = "Personal AI Employee - Digital FTE powered by Claude Code"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.163.0",
    "playwright>=1.50.0",
    "watchdog>=6.0.0",
    "python-dotenv>=1.0.1",
    "anthropic>=0.49.0",
]

[project.scripts]
filesystem-watcher = "src.watchers.filesystem_watcher:main"
gmail-watcher      = "src.watchers.gmail_watcher:main"
orchestrator       = "src.orchestrator:main"
watchdog-monitor   = "src.watchdog_monitor:run_watchdog"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
